requirements.txt
tensorflow==2.15.0
tf-keras==2.15.0

# Numpy dan OpenCV harus cocok dengan TF
numpy==1.26.0
opencv-python==4.7.0.72

# Torch stack
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# Lain-lain
facenet-pytorch>=2.5.2
retinaface-pytorch
retina-face==0.0.17
PyQt5
requests>=2.31.0
watchdog>=3.0.0
onnxruntime>=1.16.0
python-dotenv
sip